Second fix according to reviews

+++ user/src/user.service.ts
@@ -2,6 +2,7 @@ import {
   ForbiddenException,
   Injectable,
   NotFoundException,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
@@ -14,6 +15,7 @@ import {
 } from './common/dto/user.dto';
 import { JwtService } from '@nestjs/jwt';
 import { ConfigService } from '@nestjs/config';
+import { use } from 'passport';
 
 @Injectable()
 export class UserService {
@@ -23,14 +25,16 @@ export class UserService {
     @InjectModel('User') private userModel: Model<User>,
   ) {}
 
-  async createUser(params: CreateUserDto): Promise<User> {
+  async createUser(params: CreateUserDto): Promise<Omit<User, 'password'>> {
     const hashedPassword = await bcrypt.hash(params.password, 10);
     const newUser = new this.userModel({
       fullname: params.fullname,
       email: params.email,
       password: hashedPassword,
     });
-    return newUser.save();
+    const savedUser = await newUser.save();
+    const { password, ...userWithoutPass } = savedUser.toObject();
+    return userWithoutPass;
   }
 
   async login(params: LoginUserDto) {
@@ -45,22 +49,76 @@ export class UserService {
     return user;
   }
 
-  async updateUser(userId: string, params: UpdateUserDto): Promise<User> {
-    const user = await this.userModel.findByIdAndUpdate(userId, params, {
-      new: true,
-    });
+  async updateUser(
+    userId: string,
+    params: UpdateUserDto,
+  ): Promise<Omit<User, 'password'>> {
+    const user = await this.userModel.findById(userId);
     if (!user) {
       throw new NotFoundException('User not found');
     }
-    return user;
+
+    if (params.currentPassword && params.newPassword) {
+      const isMatch = await bcrypt.compare(
+        params.currentPassword,
+        user.password,
+      );
+      if (!isMatch) {
+        throw new UnauthorizedException('Current password is incorrect');
+      }
+      const hashedPassword = await bcrypt.hash(params.newPassword, 10);
+      await this.userModel.findByIdAndUpdate(userId, {
+        password: hashedPassword,
+      });
+      delete params.currentPassword;
+      delete params.newPassword;
+    }
+    const updatedUser = await this.userModel.findByIdAndUpdate(userId, params, {
+      new: true,
+    });
+    if (!updatedUser) {
+      throw new NotFoundException('User not found after update');
+    }
+    const { password, ...userWithoutPassword } = updatedUser.toObject();
+    return userWithoutPassword;
   }
 
-  async deleteUser(userId: string): Promise<string> {
+  async deleteUser(
+    userId: string,
+    requesterId: string,
+    password?: string,
+  ): Promise<string> {
+    const requester = await this.userModel.findById(requesterId);
+    if (!requester) {
+      throw new NotFoundException('Requesting user not found');
+    }
+    if (requester.role !== 'admin') {
+      await this.validateNonAdminUser(userId, password);
+    }
     const result = await this.userModel.findByIdAndDelete(userId);
     if (!result) {
-      throw new NotFoundException('User not found');
+      throw new NotFoundException('User to be deleted not found');
     }
-    return 'user deleted';
+    return 'User deleted';
+  }
+
+  private async validateNonAdminUser(
+    userId: string,
+    password?: string,
+  ): Promise<void> {
+    if (!password) {
+      throw new UnauthorizedException(
+        'Password is required for non-admin users',
+      );
+    }
+    const user = await this.userModel.findById(userId);
+    if (!user) {
+      throw new NotFoundException('User does not exist');
+    }
+    const passwordIsValid = await bcrypt.compare(password, user.password);
+    if (!passwordIsValid) {
+      throw new UnauthorizedException('Invalid password');
+    }
   }
 
   async findUserById(userId: string): Promise<User> {
